Add support for 64-bit framebuffer addresses and enhance GPU handling
- Updated framebuffer handling to use 64-bit addresses across various modules, including boot_console, framebuffer, and GPU drivers.
- Introduced atomic counters for debugging MMIO writes in the SVGA device, allowing tracking of total and text-region specific writes.
- Enhanced the GPU flush functionality to support command sending without needing a device reference.
- Improved IRQ handling with debug output for the first few interrupts.
- Added architecture-specific constants for kernel address ranges and adjusted stack validation checks accordingly.
- Refactored DLL loading to accommodate architecture differences, ensuring compatibility with both x86_64 and ARM64.
- Enhanced the kernel main function to initialize the boot console and flush the framebuffer after GPU initialization.

+++ kernel/src/drivers/gpu/virtio_gpu.rs
@@ -1070,7 +1070,7 @@ pub fn init_and_register(pci_dev: &PciDevice) -> bool {
     // This triggers the boot_console change hook which re-renders the splash
     // logo centered for the new resolution — no manual copy needed.
     crate::drivers::framebuffer::update(
-        gpu.fb_phys as u32,
+        gpu.fb_phys as u64,
         gpu.pitch,
         width,
         height,
